move Result classes into ports
