+++ src/loan-service/src/main/java/com/asc/loanservice/domain/loan/application/port/LoanEvaluationResult.java
@@ -1,4 +1,4 @@
-package com.asc.loanservice.infrastructure;
+package com.asc.loanservice.domain.loan.application.port;
 
 
 public enum LoanEvaluationResult {
