+++ src/loan-service/src/main/java/com/asc/loanservice/domain/loan/rule/port/DebtCustomerResult.java
@@ -1,4 +1,4 @@
-package com.asc.loanservice.external;
+package com.asc.loanservice.domain.loan.rule.port;
 
 
 public enum DebtCustomerResult {
