+++ src/loan-service/src/main/java/com/asc/loanservice/domain/loan/rule/port/DebtPort.java
@@ -1,7 +1,5 @@
 package com.asc.loanservice.domain.loan.rule.port;
 
-import com.asc.loanservice.external.DebtCustomerResult;
-
 
 public interface DebtPort {
